revisions to BMC Medicine
We have made substantial revisions in response to the reviewers’ constructive feedback. Specifically, we included further analyses to predict other SES factors: parental education and neighborhood disadvantage. As suggested, we revised the exclusion criteria, including head motion, resulting in a total of 6,001 participants. More details can be found in the corresponding scripts.